Add files via upload
The regression result can estimate the clinical scores which imply the concentration of biomarkers varies according to the liver condition. The designed breath collection protocol includes a brief duration exercise which is a novel approach according to the best of our knowledge in the field of liver-related study. By training our model with big dataset, a non-invasive early disease diagnosis and prognosis.